feat(examples): migrate waitForCallback tests from testing lib
