+++ packages/aws-durable-execution-sdk-js-examples/src/examples/__tests__/wait-for-callback-failing-submitter.test.ts
@@ -1,0 +1,21 @@
+import { InvocationType } from "@aws/durable-execution-sdk-js-testing";
+import { handler } from "../wait-for-callback-failing-submitter";
+import { createTests } from "./shared/test-helper";
+
+createTests({
+  name: "wait-for-callback-failing-submitter test",
+  functionName: "wait-for-callback-failing-submitter",
+  handler,
+  invocationType: InvocationType.Event,
+  tests: (runner) => {
+    // Pending resolution of https://github.com/aws/aws-durable-execution-sdk-js/issues/199
+    it.skip("should handle waitForCallback with failing submitter function errors", async () => {
+      const execution = await runner.run();
+
+      expect(execution.getResult()).toEqual({
+        success: false,
+        error: "Submitter failed",
+      });
+    });
+  },
+});
